tidy up handler and steps, but no new tests
