+++ handlers/sf-contact-merge/src/main/scala/com/gu/sf_contact_merge/validate/GetVariations.scala
@@ -1,16 +1,16 @@
 package com.gu.sf_contact_merge.validate
 
 import com.gu.sf_contact_merge.getaccounts.GetZuoraContactDetails.{EmailAddress, LastName}
-import com.gu.sf_contact_merge.validate.AssertSame.Variations
+import com.gu.sf_contact_merge.validate.GetVariations.Variations
 
-object AssertSame {
+object GetVariations {
 
   sealed trait Variations[+Element]
   case class HasNoVariations[Element](canonical: Element) extends Variations[Element]
   case class Differing[Element](variants: List[Element]) extends Variations[Element]
   case class HasAllowableVariations[Element](canonical: Element) extends Variations[Element]
 
-  def apply[Element](message: String, transform: Element => Element): AssertSame[Element] = (elements: List[Element]) => {
+  def apply[Element](message: String, transform: Element => Element): GetVariations[Element] = (elements: List[Element]) => {
     val originalVariations = elements.distinct
     val normalisedVariations = originalVariations.map(transform).distinct
     (originalVariations, normalisedVariations) match {
@@ -20,13 +20,13 @@ object AssertSame {
     }
   }
 
-  val lastName: AssertSame[LastName] = apply[LastName](
+  val forLastName: GetVariations[LastName] = apply[LastName](
     "last names",
     lastName => LastName(lastName.value.toLowerCase)
   // some seem to be entered entirely lower case, but this isn't a significant difference, so ignore
   )
 
-  val emailAddress: AssertSame[EmailAddress] = AssertSame[EmailAddress](
+  val forEmailAddress: GetVariations[EmailAddress] = GetVariations[EmailAddress](
     "emails",
     emailAddress => EmailAddress(emailAddress.value.replaceFirst("""\+gnm[0-9]?@""", "@"))
   // lowercasing is handled when it's constructed, as it's a fundamental thing (in our world) rather than just our interpretation
@@ -34,6 +34,6 @@ object AssertSame {
 
 }
 
-trait AssertSame[Element] {
+trait GetVariations[Element] {
   def apply(emailAddresses: List[Element]): Variations[Element]
 }
